% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytogeneticbandToEntrex.R
\name{cytogeneticbandToEntrex}
\alias{cytogeneticbandToEntrex}
\title{Map Cytogenetic Band to Entrez Gene IDs}
\usage{
cytogeneticbandToEntrex(cytogen_band)
}
\arguments{
\item{cytogen_band}{The cytogenetic band to map.}
}
\value{
A vector of Entrez gene IDs corresponding to the specified cytogenetic band.
}
\description{
This function maps a specified cytogenetic band to corresponding Entrez gene IDs.
}
\examples{
# Map cytogenetic band to Entrez gene IDs
result <- cytogeneticbandToEntrex("Xq28")
head(result)
}
